requirements.txt
Flask
flask[async]
python-dotenv
flake8
google.generativeai
aiohttp
asyncio